Java clients code refactoring
