+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 package-lock.json
 .DS_Store
+target
+dependency-reduced-pom.xml
