+++ SMTP/Java/smtp/gradle/settings.gradle
@@ -10,5 +10,5 @@ plugins {
     id 'org.gradle.toolchains.foojay-resolver-convention' version '0.8.0'
 }
 
-rootProject.name = 'smtp-client'
+rootProject.name = 'smtp'
 include('app')
