+++ SMTP/Java/smtp/maven/pom.xml
@@ -44,7 +44,7 @@
                             <transformers>
                                 <transformer
                                     implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.example.App</mainClass>
+                                    <mainClass>com.example.SMTP</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
